// src/types/listing.ts
export interface Listing {
  id: number;
  title: string;
  description: string;
  price: number;
  status: string;
  createdAt: string;
  viewsCount: number;
  username: string;
  siteName: string;
  categoryName: string;
  regionName: string;
  contact: string;
  themeId: number;
  imagePaths: string;
  titleEn: string;
}